+++ test/package/module_a.py
@@ -1,0 +1,1 @@
+result = "module_a"
