+++ test/package/package_a/subpackage.py
@@ -1,6 +1,9 @@
-result = 'package_a.subpackage'
+result = "package_a.subpackage"
+
+
 class PackageASubpackageObject:
     pass
+
 
 def leaf_function(a, b):
     return a + b
